+++ TEAM_1/develop/include/exec.h
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   exec.h                                             :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: hson <hson@student.42seoul.kr>             +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/03/10 13:23:45 by hson              #+#    #+#             */
+/*   Updated: 2021/03/10 14:04:54 by hson             ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef EXEC_H
 # define EXEC_H
 
@@ -25,25 +37,24 @@ int		exec_echo(t_cmd *cmd);
 void	remove_env(char *key);
 int		exec_unset(t_cmd *cmd);
 
-int		chk_arg_cnt_env();
+int		chk_arg_cnt_env(void);
 void	ft_sort_env(t_env **env, int size);
 void	print_export_sub(t_env **sort_env);
-void	print_export();
+void	print_export(void);
 void	put_env(char *str);
 int		exec_export(t_cmd *cmd);
 
 void	chk_is_cmd_path(t_cmd *cmd);
-char	**get_path_env();
+char	**get_path_env(void);
 void	free_arr(char **arr);
-char	**make_env_char();
+char	**make_env_char(void);
 void	exec_not_builtin_sub(t_cmd *cmd, char **path);
 void	exec_not_builtin(t_cmd *cmd);
 
-int		count_pipes(void);
+int		count_pipes(t_cmd *cmd);
 int		make_pipes(int fds[], int cnt);
 void	close_fds(int fds[], int pipe_count);
 void	dup_fds(int fds[], int i, int pipe_count);
-// void	wait_parent1(int fds[], pid_t pid, int cnt);
 void	wait_parent(int fds[], pid_t pid[], int cnt);
 t_cmd	*piping(t_cmd *cmd);
