+++ TEAM_1/develop/include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: hson <hson@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/23 16:11:23 by kilee             #+#    #+#             */
-/*   Updated: 2021/03/10 10:52:20 by hson             ###   ########.fr       */
+/*   Updated: 2021/03/10 13:22:46 by hson             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,6 @@
 # include "parser.h"
 # include "error.h"
 # include "exec.h"
-
 
 # define BUFFER_MAX 4096
 # define PATH_MAX 1024
