+++ TEAM_1/develop/src/exec/exec_env.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   exec_env.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: hson <hson@student.42seoul.kr>             +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/03/10 13:41:53 by hson              #+#    #+#             */
+/*   Updated: 2021/03/10 13:46:59 by hson             ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "minishell.h"
 
 int		exec_env(t_cmd *cmd)
@@ -11,9 +23,9 @@ int		exec_env(t_cmd *cmd)
 		if (env->val)
 		{
 			ft_putstr(env->key);
-			write(1,"=",1);
+			write(1, "=", 1);
 			ft_putstr(env->val);
-			write(1,"\n",1);
+			write(1, "\n", 1);
 		}
 		env = env->next;
 	}
