+++ TEAM_1/develop/src/exec/exec_export.c
@@ -1,66 +1,20 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   exec_export.c                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: hson <hson@student.42seoul.kr>             +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/03/10 13:41:12 by hson              #+#    #+#             */
+/*   Updated: 2021/03/10 13:56:34 by hson             ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "minishell.h"
 
-int		chk_arg_cnt_env()
+void	print_export(void)
 {
-	t_env *env;
-	int i;
-
-	env = g_env;
-	i = 0;
-	while (env)
-	{
-		if (env->key)
-			i++;
-		env = env->next;
-	}
-	return (i);
-}
-
-void	ft_sort_env(t_env **sort_env, int size)
-{
-	int	i;
-	int	j;
-	t_env	*tmp;
-
-	i = 0;
-	j = 0;	
-	while (j++ < size)
-	{
-		i = 0;
-		while (++i < size)
-		{
-			if(ft_strcmp(sort_env[i - 1]->key, sort_env[i]->key) > 0)
-			{
-				tmp = sort_env[i - 1];
-				sort_env[i - 1] = sort_env[i];
-				sort_env[i] = tmp;
-			}
-		}
-	}
-}
-
-void	print_export_sub(t_env **sort_env)
-{
-	int	i;
-
-	i = 0;
-	while (sort_env[i])
-	{
-		ft_putstr(sort_env[i]->key);
-		if (sort_env[i]->val)
-		{
-			write(1, "=\"", 2);
-			ft_putstr(sort_env[i]->val);
-			write(1, "\"", 1);
-		}
-		write(1, "\n", 1);
-		i++;
-	}
-}
-
-void	print_export()
-{
-	t_env 	*env;
+	t_env	*env;
 	t_env	**sort_env;
 	int		size;
 	int		i;
@@ -91,10 +45,9 @@ void	put_env(char *str)
 	while (str[i])
 	{
 		if (str[i] == '=')
-			break;
+			break ;
 		i++;
 	}
-	//printf("i = %d\n", i);
 	key = (char *)malloc(sizeof(char) * (i + 1));
 	ft_strlcpy(key, str, i + 1);
 	j = 0;
@@ -106,14 +59,13 @@ void	put_env(char *str)
 	}
 	else
 		val = 0;
-	//printf("%s : %s\n", key, val);
 	lst_change_add_env(key, val);
 }
 
 int		exec_export(t_cmd *cmd)
 {
 	t_env	*env;
-	int	i;
+	int		i;
 
 	env = g_env;
 	i = 1;
