+++ TEAM_1/develop/src/exec/exec_pwd.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   exec_pwd.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: hson <hson@student.42seoul.kr>             +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/03/10 13:42:18 by hson              #+#    #+#             */
+/*   Updated: 2021/03/10 13:50:21 by hson             ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "minishell.h"
 
 int		exec_pwd(t_cmd *cmd)
@@ -6,6 +18,6 @@ int		exec_pwd(t_cmd *cmd)
 
 	(void)cmd;
 	ft_putstr(getcwd(cwd, PATH_MAX));
-	write(1,"\n",1);
+	write(1, "\n", 1);
 	return (0);
 }
