+++ TEAM_1/develop/src/exec/pipe.c
@@ -1,35 +1,16 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   pipe.c                                             :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: hson <hson@student.42seoul.kr>             +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/03/10 13:42:30 by hson              #+#    #+#             */
+/*   Updated: 2021/03/10 14:05:07 by hson             ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "minishell.h"
-
-int		count_pipes(void)
-{
-	t_cmd	*cmd;
-	int		cnt;
-
-	cmd = g_cmd;
-	cnt = 0;
-	while (cmd)
-	{
-		if (cmd->ispipe == 1)
-			cnt++;
-		else
-			break;
-		cmd = cmd->next;
-	}
-	return (cnt);
-}
-
-int		make_pipes(int fds[], int cnt)
-{
-	int	i;
-
-	i = -1;
-	while (++i < cnt)
-	{
-		if (pipe(fds + (i * 2)) == -1)
-			ft_error();
-	}
-	return (0);
-}
 
 void	close_fds(int fds[], int pipe_count)
 {
@@ -64,12 +45,12 @@ void	wait_parent(int fds[], pid_t pid[], int cnt)
 
 t_cmd	*piping(t_cmd *cmd)
 {
-	int		fds[count_pipes() * 2];
+	int		fds[count_pipes(cmd) * 2];
 	int		cnt_pip;
 	int		i;
-	pid_t	pid[count_pipes() + 1];
+	pid_t	pid[count_pipes(cmd) + 1];
 
-	cnt_pip = count_pipes();
+	cnt_pip = count_pipes(cmd);
 	make_pipes(fds, cnt_pip);
 	i = -1;
 	while (++i < cnt_pip + 1)
